Add gitignore
The purpose of this gitignore is to ensure we are not cluttering our repository with unnecessary files from our editor environments.
As a majority of the team uses VSCode, while others use Jetbrains IDEs, this git ignore encapsulates both environments.

Credit to Github for providing gitignore templates
https://github.com/github/gitignore/blob/main/Global/JetBrains.gitignore

https://github.com/github/gitignore/blob/main/VisualStudio.gitignore

Once our tech stack is further defined, the gitignore can be altered to include other files we will not want to add.